% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{circle_line_intersections}
\alias{circle_line_intersections}
\title{Generates a set of intersection points between the cirlce and the grid lines.}
\usage{
circle_line_intersections(circle_x, circle_y, r, n_line, grid_lines)
}
\arguments{
\item{circle_x, circle_y}{The the euclidean coordinates of the center of the circle.}

\item{r}{The radius of the given circle.}

\item{n_line}{The number of grid lines.}

\item{grid_lines}{A 6 columns data frame with columns names as coor_x_1, coor_y_1, coor_x_2, coor_y_2, orient_line.
\describe{
    \item{coor_x_1, coor_y_1}{Coordinates of the left end point of the grid line }
    \item{coor_x_2, coor_y_2}{Coordinates of the right end point of the grid line }
    \item{orient_line}{Line orientation}
    \enumerate{
       \item indicates horizontal orientation
       \item indicates vetical orientation
    }
    \item{k_line}{Line numbering: bottom to top, then left to right}
}}
}
\value{
It returns a three columns data frame containing x-coordinate, y-coordanate of the intersection of the circle with the
grid, and the value of the angle betweem the x-axis and the line joining the center of the circle to the corresponding
intersection point.
}
\description{
\code{circle_line_intersections} computes the intersections points of a given circle with the grid lines along with
the angle formed with the x-axis.
}
\examples{
data(grid_line)
attach(grid_line)
circle_line_intersections(2022230,-3123109,10000,39,grid_line)
detach(grid_line)
}
