% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf_band.R
\name{Conf_Band}
\alias{Conf_Band}
\title{Visualisation of the crdible band}
\usage{
Conf_Band(OBS, samp = NULL, labl = NULL, ylbl = NULL, tobs = NULL)
}
\arguments{
\item{OBS}{Trajectory. See \code{\link{Trajectory_all}}}

\item{samp}{Tinning axis label}

\item{labl}{Axis label}

\item{t0bs}{limit of the observed time. This is used if user one want both observed and predicted band on the same graph,}
}
\value{
A 95% credible band for the obsrvation. By default it considers the removals.
}
\description{
Posterior credible band.
}
\details{
\code{Conf_Band} Provide a 95% credible band.
}
\examples{
data(rast2)
data(Dat_obs)
data(Param)
#Conf_Band(rast2,Param,Dat_obs=Dat_obs)
}
