% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_fun.R
\docType{methods}
\name{contact_fun}
\alias{contact_fun}
\title{A class for functions that may be defined in R, using pre-written native routines of Cpp functions}
\arguments{
\item{f}{A function or the name of a native routine.}

\item{PACKAGE}{optional; the name of the dynamically-loadable library in
which the native function \code{f} can be found.}

\item{proto}{optional string; a prototype against which \code{f} will be
checked.}

\item{object, x}{the \sQuote{contact_fun} object.}
}
\description{
Definition and methods of the \sQuote{contact_fun} class.
}
\details{
The \sQuote{contact_fun} class implements a common interface for user-defined procedures that can be defined in terms of R code or by compiled native routines.
}
\seealso{
\code{\link{contact}}
}
\author{
Hola K. Adrakey
}
\concept{extending the contact package}
\concept{low-level interface}
\keyword{internal}
