% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_fun.R
\docType{data}
\name{contactfunmode}
\alias{contactfunmode}
\title{A class for functions that may be defined in R, using pre-written native routines of Cpp functions}
\format{An object of class \code{list} of length 4.}
\usage{
contactfunmode
}
\description{
A class for functions that may be defined in R, using pre-written native routines of Cpp functions
}
\keyword{datasets}
