% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsimulator-package.R
\docType{package}
\name{contactsimulator-package}
\alias{contactsimulator-package}
\alias{contactsimulator,package}
\title{Inference for spatio-temporal contact type models}
\description{
The \pkg{contactsimulator} package provides facilities for inference on spatial-temporal
data using contact type distribution models.
These models are also known as state-space models, hidden Markov models, or
nonlinear stochastic dynamical systems.  One can use \pkg{contactsimulator} to fit
nonlinear, non-Gaussian dynamic models to time-series data.  The package is
both a set of tools for data analysis and a platform upon which statistical
inference methods for contacts models can be implemented.
}
\section{Data analysis using \pkg{contactsimulator}}{

\pkg{contactsimulator} provides algorithms for
\enumerate{
\item simulation of stochastic
dynamical systems; see \code{\link[=simulate-contactsimulator]{simulate}}
}
The package
also provides various tools for plotting and extracting information on
models and data.
}

\author{
Hola Adrakey
}
\concept{contactsimulator parameter estimation methods}
\concept{elementary contact methods}
\concept{information on model implementation}
