% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsimulator.R
\docType{package}
\name{contactsimulator}
\alias{contactsimulator}
\alias{contactsimulator-package}
\title{contactsimulator: A Simulator of a contact type model for spatiotemporal epidemic spread.}
\description{
The contactsimulator package provides five categories of important functions:
}
