% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_to_I}
\alias{E_to_I}
\title{Sample the latent period corresponding to each specified model.}
\usage{
E_to_I(EI_model, E, mu_lat, var_lat)
}
\arguments{
\item{EI_model}{A given integer characterising the type pf distribution for the latent period.
\enumerate{
         \item Cyclic distribution using the LER (leaf emergence rate) derived by R Allen \url{http://www.publish.csiro.au/ar/pdf/ar9780535}:
         \deqn{\frac{dL}{dt}=0.056cos(t-15) + 0.062}
         \item Gamma distribution with given rate and shape parameters: \code{\link{rgamma}}
         \item Expoential distribution with given rate: \code{\link{rexp}}, the default.

       }}

\item{E}{The time at which the contact occurs, or strictly speaking the exposure time.}

\item{mu_lat}{The mean latent period.}

\item{var_lat}{The variance of the latent period for the gamma distribution.}
}
\value{
It returns a random draw from the latent period given the time of exposure.
}
\description{
\code{E_to_I} Geratee a random draw from the distribution specified for the latent period.
}
\examples{
E_to_I(1,-20,30,10)
E_to_I(2,-20,30,10)
E_to_I(3,-20,30,10)
}
