% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{func_time_beta}
\alias{func_time_beta}
\title{Functional form of the intensity (beta) as a function of t.}
\usage{
func_time_beta(t, t_intervention, sum_beta, epsilon, omega, beta_1)
}
\arguments{
\item{t, t_intervention}{Time of the contact and the introduction of the control respectively espressed in days.}

\item{sum_beta, epsilon}{Total risk from infected premisses and the primary infection rate respectively.}

\item{omega}{Introduses the effect of seasonality in the infection rate.}
}
\value{
It returns the rate of infection in the non-homogeneous poisson process.
}
\description{
\code{func_time_beta} returns the value of the intensity as a function of t.
}
\details{
The particular case of transmission rate used here is given by \deqn{\beta(t)= n(t)\beta + \epsilon)cos^2(\omega t) if t<t_interventioin}
         where n(t) is the size of potential sources.
}
\examples{
func_time_beta(20,50,0.08,0.2,0.3)
}
\references{
\insertRef{KR08}{contactsimulator}
}
