% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{I_to_R}
\alias{I_to_R}
\title{Sample from the cyclic removal period using Brent method (Inverse tranform).}
\usage{
I_to_R(r, x_lo, x_hi, t, l, b, a)
}
\arguments{
\item{r}{The initial value.}

\item{x_lo}{A lower bound for the variable.}

\item{x_hi}{The upper bound of the variable.}

\item{t}{The current time: the infection time.}

\item{l}{A random variable in (0,1)}
}
\value{
It returns a random draw from the removal period given the time of infection for the cyclic model.
}
\description{
\code{I_to_R} Generate a random draw from the distribution specified for the latent period.
}
\examples{
I_to_R(10,0,1000,10,runif(1),0.062,0.93)
}
\references{
\insertRef{DR18}{contactsimulator}
}
