% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Inv_trans}
\alias{Inv_trans}
\title{Sample from the cyclic latent period using Brent method (Inverse tranform).}
\usage{
Inv_trans(r, x_lo, x_hi, t, l, b, a)
}
\arguments{
\item{r}{The initial value.}

\item{x_lo}{A lower bound for the variable.}

\item{x_hi}{The upper bound of the variable.}

\item{t}{The current time: the exposure time.}

\item{l}{A random variable in (0,1)}
}
\value{
It returns a random draw from the latent period given the time of exposure for the cyclic model.
}
\description{
\code{Inv_trans} Generate a random draw from the distribution specified for the latent period.
}
\examples{
Inv_trans(10,0,1000,10,runif(1))
}
\references{
\insertRef{DR18}{contactsimulator}
}
