% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Log_likelihood}
\alias{Log_likelihood}
\title{The log-likelihood of the non-homogenous Poisson process.}
\usage{
Log_likelihood(data, a, b)
}
\arguments{
\item{data}{Data frame with all records.}

\item{a, b}{Parameters of the model.}
}
\value{
It returns the log likelihood.
}
\description{
\code{Log_likelihood} returns the value of the log-likelihood.
}
\examples{
df<- data.frame(t_1=c(10,60,100),t_2=c(20,80,130),obs=c(2,3,4))
Leaf_emrgence_rate(df,0.062,0.056)
}
