% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc_leaf}
\alias{mcmc_leaf}
\title{The mcmc sampling of the non-homogenous Poisson process.}
\usage{
mcmc_leaf(data, a0, b0, samp = 1000L)
}
\arguments{
\item{data}{Data frame with all records.}

\item{a0, b0}{Innitial parameters of the model.}
}
\value{
It returns sample from the posterior distribution.
}
\description{
\code{mcmc_leaf} returns the sample from the posterior distribution.
}
\examples{
df<- data.frame(t_1=c(10,60,100),t_2=c(20,80,130))
mcmc_leaf(df,0.062,0.056)
}
