% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_IR}
\alias{r_IR}
\title{Random draw from the cyclic removal period using Brent method (Inverse tranform).}
\usage{
r_IR(n, I, a, b)
}
\arguments{
\item{n}{The sample size}

\item{I}{The time at which host become infectious, or strictly speaking the infection time.}

\item{a}{The baseline or the average infectious period.}

\item{b}{The amplitude of the removal period.}
}
\value{
It returns n random draws from the removal period given the time of infection for the cyclic model.
}
\description{
\code{r_IR} Generate a random draw from the distribution specified for the latent period.
}
\examples{
r_IR(10,10,0.062,0.93)
}
