% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rBTFinv3}
\alias{rBTFinv3}
\title{Sample the period it will take for n leaf to emerge.}
\usage{
rBTFinv3(EI_model, E, a, b, n)
}
\arguments{
\item{EI_model}{A given integer characterising the type pf distribution for the latent period.
\enumerate{
         \item Cyclic distribution using the LER (leaf emergence rate) derived by R Allen \url{http://www.publish.csiro.au/ar/pdf/ar9780535}:
         \deqn{\frac{dL}{dt}=0.056cos(t-15) + 0.062}
         \item Gamma distribution with given rate and shape parameters: \code{\link{rgamma}}
         \item Expoential distribution with given rate: \code{\link{rexp}}, the default.

       }}

\item{E}{A vector of time at which the period it will take for n leaf to emerge is computed.}

\item{a}{The baseline LER in the absence of seasonality.}

\item{b}{The amplitude of the seasonality.}

\item{n}{The number of leaf to emerge.}
}
\value{
It returns a random period of time n leaf will emerge given E.
}
\description{
\code{rBTFinv3} Generate a random draw from the distribution of leaf emergence.
}
\examples{
rBTFinv3(1,c(0,10,100),0.062,0.85,3)
}
