% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rBTFinv4}
\alias{rBTFinv4}
\title{Sample the period it took for n leaf to emerge.}
\usage{
rBTFinv4(EI_model, I, a, b, n)
}
\arguments{
\item{I}{A vector of time at which the period it took for n leaf to emerge is computed.}
}
\value{
It returns a random period of time n leaf will emerge given E.
}
\description{
\code{rBTFinv4} Generate a random draw from the distribution of leaf emergence knowing the ti.
}
\examples{
rBTFinv3(1,c(0,10,100),0.062,0.85,3)
}
