% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Samp_dis}
\alias{Samp_dis}
\title{Sample the distance at which the inoculum will travel.}
\usage{
Samp_dis(kern_model, ru, alpha1, alpha2)
}
\arguments{
\item{kern_model}{A given integer characterising the type of distribution for the kernel for both short and long range interaction.
\enumerate{
         \item exponential-exponential
         \item cauchy-cauchy
         \item exponential-cauchy
         \item cauchy
         \item exponential the default

       }}

\item{alpha1}{Dispersal scale parameter for the local spread kernel.}

\item{alpha2}{Dispersal scale parameter for the long range interaction.}

\item{ru.}{The proportion of not infected hosts in the cell containing the osurce.}
}
\value{
It returns a random distance the inoculum will travel to.
}
\description{
\code{Samp_dis} Generate a random draw from the distribution specified for the latent period.
}
\examples{
Samp_dis (1,0.5, 0.2, 0.3)
}
