% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_control_gavin.R
\name{Simulate_contact_control_LER}
\alias{Simulate_contact_control_LER}
\title{Simulation of the epidemic control process considering a distribution for the leaf}
\usage{
Simulate_contact_control_LER(f_rast = NULL, b_rast = NULL,
  farm_pos_cat = NULL, vis_int_per_cat = NULL, param, grid_lines,
  pop_grid, grid_size = 500, age_level = c(1, 1), age_dist = c(1, 0),
  m_start = 1, t_b = 1e+05, t_max = 1000, t_intervention = 1e+05,
  t_obs = 3703, EI_model = 1, kern_model = 4, rad = 1000,
  sweep_prop = c(0.5, 0.5), back_p = c(0.7, 0.5), rate_det = c(0.3,
  1), int_det = c(30, 90, 180), nb_in_b = 1, nb = 3, leav = c(3,
  6), ini = NULL)
}
\arguments{
\item{f_rast}{Population density of the grid a farm plant resides. This is filled from bottom to top, then left to right.}

\item{b_rast}{Population density of the grid a backyard plant resides. This is filled from bottom to top, then left to right.}

\item{farm_pos_cat}{A data frame of the intial conditon:
\describe{
  \item{ro}{The row at which the cell containing a farm lies in}
  \item{co}{The column at which the cell containing a farm lies in}
  \item{cat}{The category that the cell/farm belongs to}
  \item{vis_int}{Revisit interval}
  \item{tim_lst_pos}{Time of last positive}
  \item{nb_round}{Number of round of revisit}
  \item{sweep}{Decide whether to carry out the sweep over the farm: 0->no and 1->yes}
 }}

\item{vis_int_per_cat}{A data frame of the visiting intervale for the alternative strategy:
\describe{
  \item{cat}{The category that the cell/farm belongs to}
  \item{vis_int}{Revisit interval}
 }}

\item{param}{Indicating a data frame containing a vector of parameters including:
\describe{
  \item{epsion}{The primary infection rate. See \code{\link{func_time_beta}}}
  \item{beta_0}{Baseline or average transmission rate. See \code{\link{func_time_beta}}}
  \item{beta_1}{Amplitude of the seasonality. See \code{\link{func_time_beta}}}
  \item{alpha1,alpha2}{The dispersal kernel parameters.}
  \item{mu_lat,var_lat}{mean and variance of the latent period. See \code{\link{E_to_I}} for details.}
  \item{t0}{Time at which the primary source became active}.
  \item{omega}{Period of the forcing. See \code{\link{func_time_beta}}}
  \item{gama}{The mean proportion of short range dispersal events.}.
 }}

\item{pop_grid}{Population density of the grid a case resides. This is filled from bottom to top, then left to right.}

\item{grid_size}{Grid resolution
//@inheritParams circle_line_intersections}

\item{age_level}{Vectors of age level and the propportion of each age group respectively. See details.}

\item{age_dist}{Vectors of age level and the propportion of each age group respectively. See details.}

\item{m_start}{The size of initial cases. Default is 1.}

\item{t_b}{Time representing the end of the baseline programme or the start of the alternative programme}

\item{t_max}{Final observation time.}

\item{t_intervention}{Start of the intervention if any.}

\item{t_obs}{End of the observation time.}

\item{EI_model}{Take integer values to specify the type of model used for the latent period. See \code{\link{E_to_I}}}

\item{kern_model}{Take integer values to specify the type of dispersal kernel used. See \code{\link{Samp_dis}}}

\item{rad}{Sweep radius}

\item{sweep_prop}{A two element vector represention the proportion of plantation to consider for the sweep}

\item{back_p}{A two element vector represention thes Backyard assessment proportion within sweep radius:}

\item{rate_det}{Detection rate.
\describe{
  \item{1st element}{Detection rate for the grower}
  \item{2nd element}{Detection rate for the expert}
 }}

\item{int_det}{Three elements vector representing the revisit intervals:}

\item{nb_in_b}{The number of initial plants infected in category B farms}

\item{nb}{The scaling factor of backyards}

\item{leav}{The number of leaves to consider as a measurement for removal: 3 for expert to have a 100 detection}

\item{ini}{A data frame of the intial state:
\describe{
  \item{row}{The row at which the cell containing the host lies in}
  \item{col}{The column at which the cell containing the host lies in}
  \item{typ}{The type of the host (e.g. backyard 1 /plantation 0 hosts)}
  \item{x}{x-coordinate}
  \item{y}{y-coordinate}
  \item{t_e}{exposure time}
  \item{t_i}{infection time}
 }}
}
\description{
Epidemic simulation using the contact type model with the Australian control strategies.
}
\details{
\code{Simulate_contact_control_LER} provide the simulation of the epidemic process with the Australian BBTV management plan.
}
\references{
\insertRef{KR08}{contactsimulator}
\insertRef{Mee11}{contactsimulator}
}
\seealso{
\code{\link{Simulate_contact_control}}
}
