% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_NHPP_next_event}
\alias{simulate_NHPP_next_event}
\title{Simulation of the time of the next event of a non-homogeneous oisson process .}
\usage{
simulate_NHPP_next_event(t_now, t_intervention, sum_beta, epsilon, omega, t_max)
}
\arguments{
\item{t_now, t_intervention, t_max}{current time, the introduction of the control and the final observation time (usually fixed)
respectively espressed in days.}

\item{sum_beta, epsilon}{total risk from infected premisses and the primary infection rate respectively.}

\item{omega, }{introduses the effect of seasonality in the infection rate.}
}
\value{
It returns the next-event time in the non-homogeneous poisson process.
}
\description{
\code{simulate_NHPP_next_event} computes the next-event time given the current state at t.
}
\examples{
simulate_NHPP_next_event(2,50,0.08,0.2,0.3,300)
}
