% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sub_set}
\alias{Sub_set}
\title{\code{Sub_set} Count the number of observation during a certain period of time.}
\usage{
Sub_set(tr, time)
}
\arguments{
\item{tr.}{A vector giving time of events (eg removals) that occur during a time period.}

\item{time.}{A sequence of time a which observations were performed/}
}
\value{
It returns the number of observation recorded at each time.
}
\description{
\code{Sub_set} Count the number of observation during a certain period of time.
}
\examples{

time = seq(1,10)
tr = sort(runif(30,0,10))
Sub_set(tr, time)
}
