% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sum_Leaf_emergence_rate}
\alias{Sum_Leaf_emergence_rate}
\title{Functional form of the intensity of the non-homogenous Poisson process.}
\usage{
Sum_Leaf_emergence_rate(t1, t2, a, b)
}
\arguments{
\item{t1, t2}{Time of the start and end of observation respectively espressed in days.}

\item{a, b}{Parameters of the model.}
}
\value{
It returns the intensity in the non-homogeneous poisson process.
}
\description{
\code{Sum_Leaf_emergence_rate} returns the value of the intensity as a function of t1 and t2.
}
\examples{
Leaf_emrgence_rate(20,50,0.062,0.056)
}
