% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{Trajectory}
\alias{Trajectory}
\title{Temporal trajectory to computing the confidence band}
\usage{
Trajectory(epi = NULL, Dat_obs = NULL, var = "t_r")
}
\arguments{
\item{epi}{A matrix specifying the simulated epidemic. Note that the object should be of the form of the output of \code{\link{Simulate_contact_control}} or \code{\link{Simulate_contact_model}}}

\item{Dat_obs}{A 2 columns data frame wiht compoents:
\describe{
  \item{times}{ The sequence of times at which the observations are made}
  \item{obs}{the number of observation recorded at a given time}
}}

\item{var}{The observation variable. t_r is the default.}
}
\value{
A data frame of observation. By default it considers the removals.
}
\description{
Trajectory.
}
\details{
\code{Trajectory} Provide a the trjaectory of the simulated epidemic.
}
