% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{Trajectory_dis}
\alias{Trajectory_dis}
\title{Temporal trajectory to computing the wave speed}
\usage{
Trajectory_dis(epi = NULL, Dat_obs = NULL, var = "t_r")
}
\arguments{
\item{epi}{A list of matrix specifying the simulated epidemics. Note that the object should be of the form of the output of \code{\link{Simulate_contact_control}} or \code{\link{Simulate_contact_model}}}

\item{Dat_obs}{A 2 columns data frame with components:
\describe{
  \item{times}{ The sequence of times at which the observations are made}
  \item{obs}{the number of observation recorded at a given time}
}}

\item{var}{The observation variable. t_r is the default.}
}
\value{
A data frame consting of the the time points and the max distance the wave has traveled. By default it considers the removals.
}
\description{
Trajectories.
}
\details{
\code{Trajectory_dis} Provide a the trjaectories of the maximum distance the inoculum has travelled.
}
