% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{Trajectory_ripley}
\alias{Trajectory_ripley}
\title{Temporal trajectory to computing the confidence band for Ripley index}
\usage{
Trajectory_ripley(epi = NULL, Dat_obs = NULL, rast = NULL)
}
\arguments{
\item{epi}{A list of matrix specifying the simulated epidemics. Note that the object should be of the form of the output of \code{\link{Simulate_contact_control}} or \code{\link{Simulate_contact_model}}}

\item{Dat_obs}{A 2 columns data frame wiht compoents:
\describe{
  \item{times}{ The sequence of times at which the observations are made}
  \item{obs}{the number of observation recorded at a given time}
}}

\item{rast}{The raster file of the region.}

\item{var}{The observation variable. t_r is the default.}
}
\value{
A data frame consting of the the time points and the observation with the simulation. By default it considers the removals.
}
\description{
Trajectories.
}
\details{
\code{Trajectory_ripley} Provide a the trjaectories of the simulated epidemics.
}
