% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf_band.R
\name{Conf_Band}
\alias{Conf_Band}
\title{Visualisation of the crdible band}
\usage{
Conf_Band(rast, Param, age_level = c(1, 1), age_dist = c(1, 0),
  m_start = 1, t_max = 118, t_intervention = 365, EI_model = 3, Dat_obs)
}
\arguments{
\item{rast}{Raster object.}

\item{Param}{The 8 colums data frame with columns names:
\describe{
   \item{epsion}{The primary infection rate. See \code{\link{func_time_beta}}}
   \item{beta}{Rate of infection. See \code{\link{func_time_beta}}}
   \item{alpha}{The dispersal kernel parameter.}
   \item{mu_lat,var_lat}{mean and variance of the latent period. See \code{\link{E_to_I}} for details.}
   \item{omega}{Parameter characterising the seasality. See \code{\link{func_time_beta}}}
  }}

\item{age_level}{Vectors of age level and the propportion of each age group respectively. See details.}

\item{age_dist}{Vectors of age level and the propportion of each age group respectively. See details.}

\item{m_start}{The size of initial cases. Default is 1.}

\item{t_max}{Final observation time.}

\item{t_intervention}{Start of the intervention if any.}

\item{EI_model}{Take integer values to specify the type of model used for the latent period. See \code{\link{E_to_I}}}

\item{Dat_obs}{A 2 columns data frame wiht compoents:
\describe{
  \item{times}{ The sequence of times at which the observations are made}
  \item{remov}{the number of reovals recorded at a given time}
}}
}
\value{
A 95% credible band for the obsrvation. By default it considers the removals.
}
\description{
Posterior credible band.
}
\details{
\code{Conf_Band} Provide a 95% credible band.
}
\note{
The number of rows in \code{Param} corresponds to the size of the simulation expected.
}
\examples{
data(rast2)
data(Dat_obs)
data(Param)
#Conf_Band(rast2,Param,Dat_obs=Dat_obs)
}
\seealso{
\code{\link[raster]{raster}}, \code{\link{Plot_dyn}}, \code{\link{Simulate_contact_model}}, \code{\link{func_time_beta}}, \code{\link{E_to_I}}
}
